Added a tool to generate random locations easily.
Drag the TCoordTool Prefab (located in "Tools/Coordinates on Click")
onto the scene hierarchy (global), press and hold space to drag the
circle to where you want. Release space.

Everytime space is released, the circle is added to the list and the
entire list of locations is printed in JSON format to the console.